:white_check_mark: move autocomplete tests files to typescript files
