+++ tests/autocomplete/plural.ts
@@ -1,0 +1,35 @@
+import { describe, it } from 'mocha'
+import assert from 'assert'
+import { plural } from '../../src/autocomplete'
+
+describe('plural', () => {
+  it('should suggest both singular and plural forms', () => {
+    const pl = plural('test')
+    assert.deepStrictEqual([{
+      suggestion: 'est',
+      count: 1
+    }, {
+      suggestion: 'ests',
+      count: 1
+    }], pl('t'))
+  })
+
+  it('should only suggest plural form when matching with singular expression', () => {
+    const pl = plural('test')
+    assert.deepStrictEqual([{
+      suggestion: '',
+      count: 1
+    }, {
+      suggestion: 's',
+      count: 1
+    }], pl('test'))
+  })
+
+  it('should not suggest anything if expression match plural form', () => {
+    const pl = plural('story')
+    assert.deepStrictEqual([{
+      suggestion: 's',
+      count: 1
+    }], pl('storie'))
+  })
+})
